backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
Pillow==10.2.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0
aiofiles==23.2.1
numpy==1.26.4
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
